refactor: 重构 windows 活动进程获取接口和自动更新服务
